Create README.md
A visualization tool for "cymascope" simulation using RenderScript and OpenGL with a basic rendering engine. Creates a visual display coordinated to phone sound input with a variety of options of particle display, wave medium, sound source, and perspective. 

Inspired by Jenny Hans - Cymatics: A Study of Wave Phenomena and Vibration